+++ libraries/ANSI/ansi.h
@@ -2,7 +2,7 @@
 //
 //    FILE: ansi.h
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.2.0
+// VERSION: 0.2.1
 // PURPOSE: Arduino library to send ANSI escape sequences
 //    DATE: 2020-04-28
 //     URL: https://github.com/RobTillaart/ANSI
@@ -10,7 +10,7 @@
 
 #include "Arduino.h"
 
-#define ANSI_LIB_VERSION        (F("0.2.0"))
+#define ANSI_LIB_VERSION        (F("0.2.1"))
 
 
 class ANSI : public Stream
@@ -81,16 +81,18 @@ public:
   //  (for named colors, this is 25% faster than setting one then the other)
   void color(uint8_t fgcolor, uint8_t bgcolor);
 
-  //  Convert gray to ANSI 24-level gray in 4-bit colorspace
+  //  Convert gray to ANSI 24-level gray in 4-bit color space
   //  Pass in a gray level from 0 (black) to 255 (white)
   uint8_t gray2color(uint8_t gray) { return 232 + uint16_t(gray) * 24 / 256; }
   uint8_t grey2color(uint8_t grey) { return this->gray2color(grey); }
 
-  //  Convert RGB color to ANSI color in 4-bit colorspace
+  //  Convert RGB color to ANSI color in 4-bit color space
   //  Pass in a RGB level from 0 (dark) to 255 (light)
   uint8_t rgb2color(uint8_t r, uint8_t g, uint8_t b);
 
 
+  //  EXPERIMENTAL SECTION
+  //  use at own risk
   //  META
   //  deviceType is discussed
   //    - https://github.com/RobTillaart/ANSI/issues/9
@@ -101,10 +103,8 @@ public:
   int deviceType(uint32_t timeout = 100);
 
 
-  //  EXPERIMENTAL SECTION
-  //  use at own risk
   //  check if it works on your terminal
-  //                                                  TERA
+  //                                                  TERATERM
   void set132columns()   { print("\033[?3h");  };  //  +
   void set80columns()    { print("\033[?3l");  };  //  +
 
@@ -118,9 +118,9 @@ public:
   void reset()           { print("\033c");     };  //  +
 
 
-  //   NOT working on TERA TERM (or need testing)
+  //  NOT working on TERATERM (or need more testing)
   //  use at own risk
-  //  check if it works on your terminal              TERA
+  //  check if it works on your terminal              TERATERM
   /*
   void setSmoothScroll() { print("\033[?4h");  };  //  -
   void setJumpScroll()   { print("\033[?4l");  };  //  -
@@ -130,14 +130,14 @@ public:
 
   //  PRINTING
   //  use at own risk
-  //  check if it works on your terminal              TERA
+  //  check if it works on your terminal              TERATERM
   void printLine()       { print("\033[1i");   };  //  ?
   void startPrintLog()   { print("\033[4i");   };  //  ?
   void stopPrintLog()    { print("\033[5i");   };  //  ?
   */
 
 
-private:
+protected:
   size_t write(uint8_t c);
   size_t write(uint8_t * array, uint8_t length);
 
@@ -148,6 +148,17 @@ private:
 
   Stream *   _stream;
 
+};
+
+
+//////////////////////////////////////////////////////
+//
+//  DERIVED
+//
+class VT100 : public ANSI
+{
+public:
+  VT100(Stream * stream = &Serial);
 };