+++ libraries/ANSI/CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.2.1] - 2023-10-17
+- update readme.md (badges)
+- fix version in ansi.cpp
+- add a derived VT100 class (wrapper for now).
+- minor edits
+
+
 ## [0.2.0] - 2023-02-26
 - fix #13 gotoXY(column, row)
 - add ansi_gotoXY example.